Add Windows (using cygwin's rsync) support
On Windows probaly the easiest way to use rsync is to to install it
with the cygwin package. Problem with using this version is that it
requires local files to be in cygwin filesystem path to operate
correctly. It errors out with "ERROR: The source and destination
cannot both be remote." when fed with Windows path (C:\Foo). Cygwin
path ("/cygwin/c/Foo") works fine.

Also added function to normalize Windows path to use forward slashes
(this is needed regardless of rsync as Sublime API returns paths in
Windows format on Windows) and used cygpath to convert path before
it's passed to rsync.